requirements.txt
Flask==2.3.3
Flask-Mail==0.9.1
python-dotenv==1.0.0
pydantic==1.10.12
SQLAlchemy==1.4.49
email-validator==1.3.1
